requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
xgboost==2.0.3
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
nltk==3.8.1
rank-bm25==0.2.2
pydantic==2.5.0
python-multipart==0.0.6
joblib==1.3.2
tqdm==4.66.1
matplotlib==3.8.2
seaborn==0.13.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
redis==5.0.1
celery==5.3.4
